split out components into seprate files
