+++ apps/trading/pages/markets/trading-components.tsx
@@ -1,0 +1,15 @@
+export const Chart = () => <div>TODO: Chart</div>;
+export const Ticket = () => <div>TODO: Ticket</div>;
+export const Orderbook = () => <div>TODO: Orderbook</div>;
+export const Orders = () => <div>TODO: Orders</div>;
+export const Positions = () => <div>TODO: Positions</div>;
+export const Collateral = () => <div>TODO: Collateral</div>;
+
+export const Views = {
+  chart: Chart,
+  ticket: Ticket,
+  orderbook: Orderbook,
+  orders: Orders,
+  positions: Positions,
+  collateral: Collateral,
+};
